Add EmployeeContext to share store/state between components
To be able to display the employees' information in the `Dashboard` and to add new ones that can be automatically reflected in the dashboard, we need a mechanism to share such store/state. Given that we have a small application here, introducing react context seems like a good option.

This commit introduces an EmployeeContext with a value containing two properties: an array of employees and a function to update this array. The context is provided at the root level of the application so that it's accessible to all components.