Add sorting capability to table component
The dashboard's table can now be sorted by any of the columns. Bootstrap Icons library has been added to show a visual indicator about the direction of the sorting. Also an additional attribute has been added to the columns definition that it is used to show a different icon if the column contains only number values.