Preserve sort order after adding a new employee
In order to preserve the sort state of the table when adding a new employee, we need to add a new property to the EmployeeContext. This property is updated when the table triggers the callback after sorting has been executed by user.
Later on, this new property can be checked from the AddEmployeeForm and determine if sorting should be preserved or not, after that the employee is added to the context state.